Throw an exception on unknown data description types
