+++ natnet_client/__init__.py
@@ -1,6 +1,7 @@
 from .data_descriptions import *
 from .data_frame import *
 from .event import Event
+from .exceptions import NatNetError, NatNetNetworkError, NatNetProtocolError
 from .nat_net_client import NatNetClient
 from .server_info import *
 from .version import *
