+++ natnet_client/nat_net_client.py
@@ -5,26 +5,11 @@ from typing import Optional
 
 from .data_frame import DataFrame
 from .event import Event
+from .exceptions import NatNetError, NatNetNetworkError, NatNetProtocolError
 from .server_info import ServerInfo
 from .data_descriptions import DataDescriptions
 from .packet_buffer import PacketBuffer
 from .version import Version
-
-
-class NatNetError(Exception):
-    pass
-
-
-class NatNetNetworkError(NatNetError):
-    def __init__(self, socket_name: str, multicast: bool, inner_error: socket.error):
-        super().__init__(
-            f"{socket_name} socket error occurred:\n{inner_error}\nCheck Motive/Server mode requested mode agreement "
-            f"(you requested {'multi' if multicast else 'uni'}cast).")
-        self.inner_error = inner_error
-
-
-class NatNetProtocolError(NatNetError):
-    pass
 
 
 class NatNetClient:
